+++ bricks/ev3dev/modules/core.py
@@ -9,7 +9,7 @@ from pybricks.ev3devices import (
     GyroSensor,
 )
 from pybricks.parameters import Port, Stop, Direction, Button, Color
-from pybricks.tools import wait, StopWatch, DataLog
+from pybricks.tools import read_input_byte, wait, StopWatch, DataLog
 from pybricks.robotics import DriveBase
 from pybricks.media.ev3dev import ImageFile, SoundFile