+++ bricks/virtualhub/mpvarianthal.h
@@ -32,7 +32,7 @@
 
 void mp_hal_set_interrupt_char(char c);
 
-#define mp_hal_stdio_poll unused // this is not implemented, nor needed
+uintptr_t mp_hal_stdio_poll(uintptr_t);
 void mp_hal_stdio_mode_raw(void);
 void mp_hal_stdio_mode_orig(void);